refactor(hopr-transport): absorb transport/path into transport/hopr
Moves transport-path source into hopr-transport::path submodule.
Removes transport/path from the workspace.

Note: transport/tag-allocator is kept standalone (cycle risk: probe and
session depend on it, but hopr-transport depends on probe and session).
