+++ Cargo.toml
@@ -15,7 +15,6 @@ members = [
   "protocols/start",
   "transport/hopr",
   "transport/mixer",
-  "transport/path",
   "transport/probe",
   "transport/session",
   "transport/tag-allocator",
@@ -178,7 +177,6 @@ hopr-protocol-start = { path = "protocols/start", default-features = false }
 hopr-ticket-manager = { path = "impls/ticket-manager" }
 hopr-transport = { path = "transport/hopr", default-features = false }
 hopr-transport-mixer = { path = "transport/mixer", default-features = false }
-hopr-transport-path = { path = "transport/path", default-features = false }
 hopr-transport-probe = { path = "transport/probe", default-features = false }
 hopr-transport-session = { path = "transport/session", default-features = false }
 hopr-transport-tag-allocator = { path = "transport/tag-allocator", default-features = false }
