+++ transport/hopr/src/path/errors.rs
